gave each of the classes in graph.cs its own file
