+++ Editor/Eazy Window/Data Nodes.meta
