+++ Editor/Eazy Window/Data Nodes/BigTestNode.cs
