+++ Editor/Eazy Window/Data Nodes/TestNode.cs.meta
