+++ Editor/Eazy Window/Serializer/Graph.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fcf361fe52709c94eb5fdd889edaa49c
+guid: 0b62b376f0a6264418bdc210765d2e7c
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
