+++ Editor/Eazy Window/Serializer/Graph/SaveConnection.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d0b5312236f1cd94586d6662d609cef8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
