+++ Editor/Eazy Window/Serializer/Graph/SaveNode.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d1c1e94f2d4cd2c4fa95dc6723b971be
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
