+++ Editor/Eazy Window/Serializer/Serialization/Scriptable Object/Data Objects/ConnectionData.cs.meta
