+++ Editor/Eazy Window/Serializer/Serialization/Scriptable Object/Data Objects/GraphData.cs
