+++ Editor/Eazy Window/Serializer/Serialization/Scriptable Object/Data Objects/NodeData.cs
